+++ src/components/atoms/tag/index.js
@@ -1,0 +1,4 @@
+import Tag, { StyledTag } from './tag'
+
+export default Tag
+export { StyledTag }